Make coupled and atomic modules a part of component
