+++ src/component/atomic.rs
