+++ src/lib.rs
@@ -5,9 +5,7 @@ extern crate self as xdevs;
 #[cfg(feature = "std")]
 extern crate std;
 
-pub mod atomic;
 pub mod component;
-pub mod coupled;
 #[cfg(feature = "alloc")]
 pub mod devstone;
 pub mod export;
@@ -19,9 +17,7 @@ pub mod rt_engine;
 pub mod simulator;
 pub mod traits;
 
-pub use atomic::Atomic;
-pub use component::{AtomicKind, Component, CoupledKind};
-pub use coupled::Coupled;
+pub use component::{atomic::Atomic, coupled::Coupled, AtomicKind, Component, CoupledKind};
 pub use embassy_time::{Duration, Instant};
 pub use port::Port;
 pub use simulator::{Config, Simulator};
